requirements.txt
oletools >= 0.60
python-magic >= 0.4.24
matplotlib
numpy
python-whois
nslookup
